agv/src/main/java/com/bobbyprod/agv/AgvService.java
package com.bobbyprod.agv;

import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.stereotype.Service;

@Service
@EnableConfigurationProperties(ServiceProperties.class)
public class AgvService {
    private final ServiceProperties serviceProperties;

    public AgvService(ServiceProperties serviceProperties) {
        this.serviceProperties = serviceProperties;
    }

    public String message() {
        return this.serviceProperties.getMessage();
    }
}
